Update staking.md
Adding a note to double check they have voted or proxied as that will be a common error that doesn't make much sense.

Example errors from bloks and anchor that had me confused:

```
Error
Cannot read properties of undefined (reading 'id')

Error
An invalid request was sent in, please check the nested errors for details. ()

{"type":"invalid_input","message":"An invalid request was sent in, please check the nested errors for details.","fields":[{"name":"actor","value":"2serxzaeb1ry","error":"Account has not voted and has not proxied."}]}
```

I had to go to the developer tools to find out what was actually going on, so it was a bit confusing.